requirements.txt
Django~=4.0.4
scikit-image~=0.19.2
tensorflow~=2.8.0
djangorestframework~=3.13.1
opencv-python~=4.5.5.64
numpy~=1.22.3
matplotlib~=3.5.1
cffi~=1.14.5
future~=0.16.0
Pillow~=9.1.0
ipython~=8.2.0
pip~=21.0.1
attrs~=20.3.0
zipp~=3.8.0
toml~=0.10.2
wheel~=0.36.2
contrib~=0.3.0
cloudbiolinux~=0.1
cryptography~=3.4.6
Jinja2~=2.10.3
rsa~=4.8
pyasn1~=0.4.8
six~=1.16.0
gast~=0.5.3
dataset~=1.5.2
h5py~=3.6.0
parso~=0.8.3
jedi~=0.18.1
colorama~=0.4.4
pytz~=2022.1
scipy~=1.8.0
keras~=2.8.0
pydotplus~=2.0.2
pandas~=1.2.3
wrapt~=1.14.0
requests~=2.27.1
Cython~=0.29.28
setuptools~=41.2.0
sqlparse~=0.4.2
asgiref~=3.5.0
PyYAML~=6.0
Geometry~=0.0.23
point~=0.0.1
polygon~=0.9.6
cachetools~=5.0.0
certifi~=2021.10.8
urllib3~=1.26.9
imgaug~=0.4.0
imageio~=2.18.0
tifffile~=2022.4.22
Pygments~=2.12.0
traitlets~=5.1.1
decorator~=5.1.1
packaging~=21.3
backcall~=0.2.0
pickleshare~=0.7.5
pexpect~=4.8.0
wcwidth~=0.2.5
ptyprocess~=0.7.0
Shapely~=1.8.1.post1
networkx~=2.8
sklearn~=0.0
scikit-learn~=0.24.1
python-dateutil~=2.8.2
Markdown~=3.3.6
oauthlib~=3.2.0
idna~=3.3
chardet~=4.0.0
Werkzeug~=2.1.1
asttokens~=2.0.5
executing~=0.8.3
fonttools~=4.33.2
path~=15.1.0
pyparsing~=3.0.8
cycler~=0.11.0
kiwisolver~=1.4.2
tensorboard~=2.8.0
flatbuffers~=2.0
astunparse~=1.6.3
termcolor~=1.1.0
pycocotools~=2.0.4
botocore~=1.19.63
pathlib2~=2.3.7.post0
click~=7.1.2
open3d~=0.14.1